feat(bootstrap): split bootstrap from the core
