+++ demo/src/formly/wrappers/horizontal.wrapper.ts
@@ -1,5 +1,5 @@
 import { Component, ViewContainerRef, ViewChild } from '@angular/core';
-import { FieldWrapper } from 'ng-formly';
+import { FieldWrapper } from 'ng-formly/core';
 @Component({
   selector: 'formly-wrapper-horizontal',
   template: `
