+++ src/core/core.module.ts
@@ -2,34 +2,13 @@ import { NgModule, ModuleWithProviders, ANALYZE_FOR_ENTRY_COMPONENTS } from '@an
 import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 import { FormlyForm } from './components/formly.form';
-import { FormlyFieldConfig } from './components/formly.field.config';
 import { FormlyField } from './components/formly.field';
 import { FormlyAttributes } from './components/formly.attributes';
 import { FormlyConfig, ConfigOption, FORMLY_CONFIG_TOKEN } from './services/formly.config';
 import { FormlyFormBuilder } from './services/formly.form.builder';
 import { FormlyValidationMessages } from './services/formly.validation-messages';
-import { FormlyPubSub, FormlyEventEmitter } from './services/formly.event.emitter';
-import { Field } from './templates/field';
-import { FieldType } from './templates/field.type';
-import { FieldWrapper } from './templates/field.wrapper';
+import { FormlyPubSub } from './services/formly.event.emitter';
 import { FormlyGroup } from './components/formly.group';
-
-export {
-  ConfigOption,
-  FormlyAttributes,
-  FormlyFormBuilder,
-  FormlyField,
-  FormlyFieldConfig,
-  FormlyForm,
-  FormlyConfig,
-  FormlyPubSub,
-  FormlyValidationMessages,
-  FormlyEventEmitter,
-
-  Field,
-  FieldType,
-  FieldWrapper,
-};
 
 const FORMLY_DIRECTIVES = [FormlyForm, FormlyField, FormlyAttributes, FormlyGroup];
