+++ src/ui-bootstrap/types/multicheckbox.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { FormGroup, FormControl, AbstractControl } from '@angular/forms';
-import { FieldType, FormlyFieldConfig } from '../../core/core';
+import { FieldType, FormlyFieldConfig } from '../../core';
 
 @Component({
   selector: 'formly-field-multicheckbox',
