+++ src/ui-bootstrap/types/types.ts
@@ -1,15 +1,6 @@
-import { FormlyFieldCheckbox } from './checkbox';
-import { FormlyFieldMultiCheckbox } from './multicheckbox';
-import { FormlyFieldInput } from './input';
-import { FormlyFieldRadio } from './radio';
-import { FormlyFieldTextArea } from './textarea';
-import { FormlyFieldSelect } from './select';
-
-export {
-  FormlyFieldCheckbox,
-  FormlyFieldMultiCheckbox,
-  FormlyFieldInput,
-  FormlyFieldRadio,
-  FormlyFieldTextArea,
-  FormlyFieldSelect,
-};
+export { FormlyFieldCheckbox } from './checkbox';
+export { FormlyFieldMultiCheckbox } from './multicheckbox';
+export { FormlyFieldInput } from './input';
+export { FormlyFieldRadio } from './radio';
+export { FormlyFieldTextArea } from './textarea';
+export { FormlyFieldSelect } from './select';
