+++ src/ui-bootstrap/wrappers/wrappers.ts
@@ -1,11 +1,4 @@
-import { FormlyWrapperFieldset } from './fieldset';
-import { FormlyWrapperLabel } from './label';
-import { FormlyWrapperDescription } from './description';
-import { FormlyWrapperValidationMessages } from './message-validation';
-
-export {
-  FormlyWrapperFieldset,
-  FormlyWrapperLabel,
-  FormlyWrapperDescription,
-  FormlyWrapperValidationMessages,
-};
+export { FormlyWrapperFieldset } from './fieldset';
+export { FormlyWrapperLabel } from './label';
+export { FormlyWrapperDescription } from './description';
+export { FormlyWrapperValidationMessages } from './message-validation';
